updated console.py
