--- file.json
@@ -1,1 +1,0 @@
-{"BaseModel.68401c5d-7124-4dbc-a338-7aa8756d13e8": {"__class__": "BaseModel", "id": "68401c5d-7124-4dbc-a338-7aa8756d13e8", "created_at": "2023-05-11T12:08:59.319328", "updated_at": "2023-05-11T12:08:59.319345"}, "BaseModel.33f6e24e-9f9f-4998-882d-954863cf50da": {"__class__": "BaseModel", "id": "33f6e24e-9f9f-4998-882d-954863cf50da", "created_at": "2023-05-11T12:09:14.369402", "updated_at": "2023-05-11T12:09:14.369414"}}
