+++ models/state.py
@@ -2,6 +2,7 @@
 
 """ This module holds the state class """
 
+from models.base_model import BaseModel
 
 class State(BaseModel):
     """ This class holds all state objects """
